Fixing functional test
