+++ x-pack/test/alerting_api_integration/security_and_spaces/tests/telemetry/alerting_telemetry.ts
@@ -11,11 +11,11 @@ import {
   getUrlPrefix,
   getEventLog,
   getTestRuleData,
-  ObjectRemover,
   TaskManagerDoc,
   ESTestIndexTool,
 } from '../../../common/lib';
 import { FtrProviderContext } from '../../../common/ftr_provider_context';
+import { setupSpacesAndUsers } from '..';
 
 // eslint-disable-next-line import/no-default-export
 export default function createAlertingTelemetryTests({ getService }: FtrProviderContext) {
@@ -24,26 +24,18 @@ export default function createAlertingTelemetryTests({ getService }: FtrProvider
   const retry = getService('retry');
   const supertestWithoutAuth = getService('supertestWithoutAuth');
   const esTestIndexTool = new ESTestIndexTool(es, retry);
+  const esArchiver = getService('esArchiver');
 
   describe('alerting telemetry', () => {
     const alwaysFiringRuleId: { [key: string]: string } = {};
-    const objectRemover = new ObjectRemover(supertest);
 
     before(async () => {
-      // reset the state in the telemetry task
-      await es.update({
-        id: `task:Alerting-alerting_telemetry`,
-        index: '.kibana_task_manager',
-        body: {
-          doc: {
-            task: {
-              state: '{}',
-            },
-          },
-        },
-      });
+      await esArchiver.load('x-pack/test/functional/es_archives/event_log_telemetry');
+      await setupSpacesAndUsers(getService);
     });
-    after(() => objectRemover.removeAll());
+    after(async () => {
+      await esArchiver.unload('x-pack/test/functional/es_archives/event_log_telemetry');
+    });
 
     beforeEach(async () => {
       await esTestIndexTool.destroy();
@@ -63,7 +55,6 @@ export default function createAlertingTelemetryTests({ getService }: FtrProvider
           secrets: {},
         })
         .expect(200);
-      objectRemover.add(space, createdConnector.id, 'action', 'actions');
       return createdConnector.id;
     }
 
@@ -75,7 +66,6 @@ export default function createAlertingTelemetryTests({ getService }: FtrProvider
         .auth(Superuser.username, Superuser.password)
         .send(getTestRuleData(ruleOverwrites));
       expect(ruleResponse.status).to.eql(200);
-      objectRemover.add(space, ruleResponse.body.id, 'rule', 'alerting');
       return ruleResponse.body.id;
     }
